feat: remove XModuleMixin legacy attrs from all xblocks
