+++ xblocks_contrib/video/video.py
@@ -29,25 +29,25 @@ from lxml import etree
 from opaque_keys.edx.keys import CourseKey
 from opaque_keys.edx.locator import AssetLocator
 from web_fragments.fragment import Fragment
+from webob import Response
+from webob.multidict import MultiDict
 from xblock.completable import XBlockCompletionMode
-from xblock.core import Scope, XBlock
-from xblock.fields import ScopeIds, UserScope
+from xblock.core import XBlock
+from xblock.fields import ScopeIds
 from xblock.utils.resources import ResourceLoader
 
-from xblocks_contrib.common.xml_utils import (
+from xblocks_contrib.legacy_utils.xml_utils import (
     EdxJSONEncoder,
     LegacyXmlMixin,
     deserialize_field,
     is_pointer_tag,
     name_to_pathname,
 )
-from xblocks_contrib.video.ajax_handler_mixin import AjaxHandlerMixin
 from xblocks_contrib.video.bumper_utils import bumperize
 from xblocks_contrib.video.cache_utils import request_cached
 from xblocks_contrib.video.constants import ATTR_KEY_REQUEST_COUNTRY_CODE, ATTR_KEY_USER_ID, PUBLIC_VIEW, STUDENT_VIEW
 from xblocks_contrib.video.exceptions import TranscriptNotFoundError
 from xblocks_contrib.video.mixin import LicenseMixin
-from xblocks_contrib.video.studio_metadata_mixin import StudioMetadataMixin
 from xblocks_contrib.video.validation import StudioValidation, StudioValidationMessage
 from xblocks_contrib.video.video_handlers import VideoStudentViewHandlers, VideoStudioViewHandlers
 from xblocks_contrib.video.video_static_content_utils import (
@@ -60,6 +60,7 @@ from xblocks_contrib.video.video_transcripts_utils import (
     VideoTranscriptsMixin,
     clean_video_id,
     get_endonym_or_label,
+    get_html5_ids,
     subs_filename,
 )
 from xblocks_contrib.video.video_utils import (
@@ -110,9 +111,7 @@ EXPORT_IMPORT_STATIC_DIR = 'static'
 @XBlock.needs('i18n', 'user')
 class VideoBlock(
         VideoFields, VideoTranscriptsMixin, VideoStudioViewHandlers, VideoStudentViewHandlers,
-        LegacyXmlMixin, XBlock,
-        AjaxHandlerMixin, StudioMetadataMixin,
-        LicenseMixin):
+        LegacyXmlMixin, XBlock, LicenseMixin):
     """
     XML source example::
 
@@ -149,28 +148,6 @@ class VideoBlock(
     js_module_name = "TabsEditingDescriptor"
 
     uses_xmodule_styles_setup = True
-
-    @property
-    def display_name_with_default(self):
-        """
-        Return a display name for the module: use display_name if defined in
-        metadata, otherwise convert the url name.
-        """
-        return (
-            self.display_name if self.display_name is not None
-            else self.usage_key.block_id.replace('_', ' ')
-        )
-
-    @property
-    def xblock_kvs(self):
-        """
-        Retrieves the internal KeyValueStore for this XModule.
-
-        Should only be used by the persistence layer. Use with caution.
-        """
-        # if caller wants kvs, caller's assuming it's up to date; so, decache it
-        self.save()
-        return self._field_data._kvs  # pylint: disable=protected-access
 
     def get_transcripts_for_student(self, transcripts, dest_lang=None):
         """Return transcript information necessary for rendering the XModule student view.
@@ -504,7 +481,7 @@ class VideoBlock(
             'bumper_metadata': json.dumps(self.bumper['metadata']),  # pylint: disable=E1101
             'cdn_eval': cdn_eval,
             'cdn_exp_group': cdn_exp_group,
-            'display_name': self.display_name_with_default,
+            'display_name': self.display_name_with_default,  # pylint: disable=no-member
             'download_video_link': download_video_link,
             'is_video_from_same_origin': is_video_from_same_origin,
             'handout': self.handout,
@@ -1189,58 +1166,44 @@ class VideoBlock(
             )
         return None
 
-    def bind_for_student(self, user_id, wrappers=None):
+    @property
+    def ajax_url(self):
         """
-        Set up this XBlock to act as an XModule instead of an XModuleDescriptor.
-
-        Arguments:
-            user_id: The user_id to set in scope_ids
-            wrappers: These are a list functions that put a wrapper, such as
-                      LmsFieldData or OverrideFieldData, around the field_data.
-                      Note that the functions will be applied in the order in
-                      which they're listed. So [f1, f2] -> f2(f1(field_data))
+        Returns the URL for the ajax handler.
         """
+        return self.runtime.handler_url(self, 'ajax_handler', '', '').rstrip('/?')
 
-        # Skip rebinding if we're already bound a user, and it's this user.
-        if self.scope_ids.user_id is not None and user_id == self.scope_ids.user_id:
-            if getattr(self.runtime, "position", None):
-                # update the position of the tab
-                self.position = self.runtime.position  # pylint: disable=attribute-defined-outside-init
-            return
+    @XBlock.handler
+    def ajax_handler(self, request, suffix=None):
+        """
+        XBlock handler that wraps `ajax_handler`
+        """
+        class FileObjForWebobFiles:
+            """
+            Turn Webob cgi.FieldStorage uploaded files into pure file objects.
 
-        # If we are switching users mid-request, save the data from the old user.
-        self.save()
+            Webob represents uploaded files as cgi.FieldStorage objects, which
+            have a .file attribute.  We wrap the FieldStorage object, delegating
+            attribute access to the .file attribute.  But the files have no
+            name, so we carry the FieldStorage .filename attribute as the .name.
 
-        # Update scope_ids to point to the new user.
-        self.scope_ids = self.scope_ids._replace(user_id=user_id)
+            """
+            def __init__(self, webob_file):
+                self.file = webob_file.file
+                self.name = webob_file.filename
 
-        # Clear out any cached instantiated children.
-        self.clear_child_cache()
+            def __getattr__(self, name):
+                return getattr(self.file, name)
 
-        # Clear out any cached field data scoped to the old user.
-        for field in self.fields.values():
-            if field.scope in (Scope.parent, Scope.children):
-                continue
+        # WebOb requests have multiple entries for uploaded files.  handle_ajax
+        # expects a single entry as a list.
+        request_post = MultiDict(request.POST)
+        for key in set(request.POST.keys()):
+            if hasattr(request.POST[key], "file"):
+                request_post[key] = list(map(FileObjForWebobFiles, request.POST.getall(key)))
 
-            if field.scope.user == UserScope.ONE:
-                field._del_cached_value(self)  # pylint: disable=protected-access
-                # not the most elegant way of doing this, but if we're removing
-                # a field from the module's field_data_cache, we should also
-                # remove it from its _dirty_fields
-                if field in self._dirty_fields:
-                    del self._dirty_fields[field]
-
-        if wrappers:
-            # Put user-specific wrappers around the field-data service for this block.
-            # Note that these are different from modulestore.xblock_field_data_wrappers, which are not user-specific.
-            wrapped_field_data = self.runtime.service(self, "field-data-unbound")
-            for wrapper in wrappers:
-                wrapped_field_data = wrapper(wrapped_field_data)
-            self._bound_field_data = wrapped_field_data  # pylint: disable=attribute-defined-outside-init
-            if getattr(self.runtime, "uses_deprecated_field_data", False):
-                # This approach is deprecated but OldModuleStoreRuntime still requires it.
-                # For SplitModuleStoreRuntime, don't set ._field_data this way.
-                self._field_data = wrapped_field_data
+        response_data = self.handle_ajax(suffix, request_post)
+        return Response(response_data, content_type='application/json', charset='UTF-8')
 
     @classmethod
     def definition_from_xml(cls, xml_object, system):
@@ -1248,20 +1211,66 @@ class VideoBlock(
             return {"data": ""}, []
         return {"data": etree.tostring(xml_object, pretty_print=True, encoding="unicode")}, []
 
-    def get_explicitly_set_fields_by_scope(self, scope=Scope.content):
+    @property
+    def editable_metadata_fields(self):
         """
-        Get a dictionary of the fields for the given scope which are set explicitly on this xblock. (Including
-        any set to None.)
+        Returns the metadata fields to be edited in Studio.
         """
-        result = {}
-        for field in self.fields.values():
-            if field.scope == scope and field.is_set_on(self):
+        editable_fields = super().editable_metadata_fields  # pylint: disable=no-member
+
+        settings_service = self.runtime.service(self, 'settings')
+        if settings_service:
+            xb_settings = settings_service.get_settings_bucket(self)
+            if not xb_settings.get("licensing_enabled", False) and "license" in editable_fields:
+                del editable_fields["license"]
+
+        # Default Timed Transcript a.k.a `sub` has been deprecated and end users shall
+        # not be able to modify it.
+        editable_fields.pop('sub')
+
+        languages = [{'label': label, 'code': lang} for lang, label in settings.ALL_LANGUAGES]
+        languages.sort(key=lambda lang_item: lang_item['label'])
+        editable_fields['transcripts']['custom'] = True
+        editable_fields['transcripts']['languages'] = languages
+        editable_fields['transcripts']['type'] = 'VideoTranslations'
+
+        # We need to send ajax requests to show transcript status
+        # whenever edx_video_id changes on frontend. Thats why we
+        # are changing type to `VideoID` so that a specific
+        # Backbonjs view can handle it.
+        editable_fields['edx_video_id']['type'] = 'VideoID'
+
+        # `public_access` is a boolean field and by default backbonejs code render it as a dropdown with 2 options
+        # but in our case we also need to show an input field with dropdown, the input field will show the url to
+        # be shared with leaners. This is not possible with default rendering logic in backbonjs code, that is why
+        # we are setting a new type and then do a custom rendering in backbonejs code to render the desired UI.
+        editable_fields['public_access']['type'] = 'PublicAccess'
+        editable_fields['public_access']['url'] = self.get_public_video_url()
+
+        # construct transcripts info and also find if `en` subs exist
+        transcripts_info = self.get_transcripts_info()
+        possible_sub_ids = [self.sub, self.youtube_id_1_0] + get_html5_ids(self.html5_sources)
+        video_config_service = self.runtime.service(self, 'video_config')
+        if video_config_service:
+            for sub_id in possible_sub_ids:
                 try:
-                    result[field.name] = field.read_json(self)
-                except TypeError as exception:
-                    exception_message = f"{exception}, Block-location:{self.usage_key}, Field-name:{field.name}"
-                    raise TypeError(exception_message) from exception
-        return result
+                    _, sub_id, _ = video_config_service.get_transcript(
+                        self, lang='en', output_format=TranscriptExtensions.TXT
+                    )
+                    transcripts_info['transcripts'] = dict(transcripts_info['transcripts'], en=sub_id)
+                    break
+                except TranscriptNotFoundError:
+                    continue
+
+        editable_fields['transcripts']['value'] = transcripts_info['transcripts']
+        editable_fields['transcripts']['urlRoot'] = self.runtime.handler_url(
+            self,
+            'studio_transcript',
+            'translation'
+        ).rstrip('/?')
+        editable_fields['handout']['type'] = 'FileUploader'
+
+        return editable_fields
 
     @staticmethod
     def workbench_scenarios():
